# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/perfil": {
      "get": {
        "operationId": "AuthController_getPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil de usuario y finca",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar el perfil del usuario autenticado",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/invitar": {
      "post": {
        "operationId": "AuthController_invitarUsuario",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/confirmar-invitacion": {
      "post": {
        "operationId": "AuthController_confirmarInvitacion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/catalogos/razas": {
      "get": {
        "description": "Devuelve las razas globales del sistema más las propias de la finca. Cada raza trae sus días de gestación, que es lo que MOD-03 usa para calcular la FPP.",
        "operationId": "CatalogosController_findAllRazas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Listado de razas ordenado por nombre"
          },
          "401": {
            "description": "Token ausente, inválido o sin claims de tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Catálogo de razas disponible para la finca",
        "tags": [
          "Catálogos"
        ]
      }
    },
...
    "/tratamientos": {
      "post": {
        "operationId": "TratamientosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTratamientoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tratamientos Sanitarios"
        ]
      }
    },
    "/tratamientos/animal/{animalId}/estado-sanitario": {
      "get": {
        "operationId": "TratamientosController_getEstadoSanitario",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tratamientos Sanitarios"
        ]
      }
    },
    "/tratamientos/animal/{animalId}": {
      "get": {
        "operationId": "TratamientosController_findAllByAnimal",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tratamientos Sanitarios"
        ]
      }
    },
    "/tratamientos/{id}": {
      "put": {
        "operationId": "TratamientosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTratamientoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tratamientos Sanitarios"
        ]
      }
    },
    "/catalogos/medicamentos": {
      "get": {
        "operationId": "SanitaryController_getMedicamentos",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Token ausente, inválido o sin claims de tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener catálogo de medicamentos veterinarios del tenant",
        "tags": [
          "Catálogos Sanitarios"
        ]
      }
    },
    "/catalogos/padecimientos": {
      "get": {
        "operationId": "SanitaryController_getPadecimientos",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Token ausente, inválido o sin claims de tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener catálogo de padecimientos del tenant con fármaco sugerido",
        "tags": [
          "Catálogos Sanitarios"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResDigital API",
    "description": "API del sistema de gestión ganadera ResDigital",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "rol": {
            "type": "string",
            "enum": [
              "propietario",
              "administrador",
              "peon",
              "veterinario"
            ]
          },
          "nombreCompleto": {
            "type": "string"
          },
          "correo": {
            "type": "string",
            "format": "email"
          },
          "nombreFinca": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "tenantId",
          "rol",
          "nombreCompleto",
          "correo",
          "nombreFinca"
        ]
      },
      "InviteUserDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmInvitationDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTratamientoDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTratamientoDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}